Fix currency representation and formatting
To prevent problems with floating point accuracy, we represent prices
in cents internally, until they are formatted at the domain boundary,
for which we use BigDecimal.
We format prices in the official format with currency symbol.

1. Currency data type:

An alternative solution would be to use BigDecimal everywhere, but in
this case we opt against it for the following reasons:
- BigDecimals in itself are not a currency type, but an accurate
  fixed point decimal type.
- BigDecimals are mutable and not constrained to a specific decimal
  point position, so an illegal currency value like `14.999` could
  be accidentally created and passed around through the system before
  it is realized, causing potentially difficult-to-debug errors (besides
  the standard mutability issues).
- If no fractional cents are required there is no need for the ability
  to move the decimal point further than 2 positions, i.e. creating
  more than 1 fixed point representation.

In general, I would recommend a domain type for money (e.g. a `Money` value
object), which uses either long or BigDecimal internally, but without
leaking that implementation detail to the outside, allowing to easily
replace long with BigDecimal when needed without affecting the users of
the type. In this case, the simple solution of using prices in cents, to fix
the floating point accurracy problem, seems good enough, since the focus of
the kata is on duplication/DRY. However, a domain type for money should
probably be considered in other design focused katas (e.g. Information Hiding,
YAGNI, etc.).

Further reading on currency representation:
https://stackoverflow.com/questions/3730019/why-not-use-double-or-float-to-represent-currency
https://stackoverflow.com/a/45406786/3726133
https://stackoverflow.com/a/45222775/3726133

2. Currency Formatting:

Euro amounts are formatted in a specific, standardized way, which is the
same in most countries, e.g., "14,99 €".

Something that is often overlooked by programmers is that most unit symbols
are separated from the value by a non-breaking space, which prevents the
case of, e.g., "14,99
€" being split into 2 lines, and instead moves the whole formatted value
to the next line if it does not fit. Standard libraries typically do that by
default (including, Java's `NumberFormat` for currency values. So we only add
it manually in cases where we format manually and in tests.
The non-breaking space has the hex value `00A0` in unicode, which is written
in a Java string literal as `"\u00A0"`.

Further reading on currency formatting:
https://en.wikipedia.org/wiki/Language_and_the_euro
https://stackoverflow.com/questions/12560629/java-converting-long-into-currency
https://en.wikipedia.org/wiki/Non-breaking_space