Solution example: Task 1 and 2
Task 1
======

a)

The FileSystemUserRepository implies details about its implementation.
This can cause its callers to implicitly depend on those implementation
details, because as soon as they know about them, it is difficult to
ignore them. The `FileSystemUserRepository` interface is way too specific.
As the consumer of the interface, we don't care about the details of the
persistence mechanisms in use.

We can also move the interface to the domain layer. Accessing domain
entities is a domain concern, and consumers on lower levels (e.g.,
use cases and services) should only depend on the higher levels (in
this case the domain, i.e., the business rules).

b)

At the same time, the `FileSystemUserRepositoryImpl` has to add an
`Impl` suffix, which arguably decreases readability, and only does
so because the name has to be unique.

This leads to Indirection Without Abstraction. By fixing the name
(`UserRepository`) we can get rid of the ugly `Impl` suffix. The
fact that a class implements an interface is an implementation
detail, and in many cases, classes can (and should) implement
multiple interfaces anyway. This also makes it more obvious that
`CsvFileUserRepository` is even clearer - on the implementation
side, not the interface side.

c)

In the case of the `ProductRepository` interface and `InMemoryProductRepository`
implementation, we already have the situation that we want to get to in
the case of the `UserRepository`(see also the diagram below for Task 2).
In summary:
- The interface does not imply anything about the implementation mechanism.
- The implementation name is more specific than the interface, i.e., note
  how we could easily use a different persistence mechanism instead, for
  example, an `SQLiteProductRepository`.
- The consumer of the `ProductRepository` interface, i.e., the `GetUserInfoUseCase`,
  accesses `Product` instances via the interface, which resides in the
  domain layer. That way, the use case only depends on the higher-level
  domain layer, and is persistence-agnostic. The repository implementation
  is injected into the use case like a plug-in.

d)

The `ReceiptPresenter` and `UserInfoPresenter` define the interfaces
that they intend to consume. This is useful if there could be multiple
possible implementations (or at least we do not care about the specifics
of the implementation in the consumer), but only one consumer of the
interface. This is basically a plug-in interface, and the plug-in can
choose to use the same name, without an `Impl` suffix, because the
interface name is fully qualified via its consumer's class name.
We do not need to fix anything here, because it is one of many viable
ways of implementing inward-facing dependencies.

Task 2
======

In Task 2, we refactor to solve the issues identified in Task 1 a) and b):
We move the interface to the right layer, and we make the interface name
more abstract and the implementation name more concrete.

Before:
                       ||
  GetUserInfoUseCase --||--> FileSystemUserRepository <---- FileSystemUserRepositoryImpl
               |       ||                                      ^
               |       ||                                      |
               |-------||---- *implicit dependency*------------|
                       ||
                       ||
                       || *layer boundary*

After:
                       ||                      ||
  GetUserInfoUseCase --||--> UserRepository <--||-- CsvFileUserRepository
                       ||                      ||
                       ||                      ||
                       ||  *layer boundaries*  ||