Solution example: Task 3
We rename `GetUserInfo...` to `GetFullName...` because it describes
the intent more precisely.

This is not the only way so resolve this problem of implicit dependencies
between abstraction levels. We could also rename the endpoint, but that
has consequences for the outside world, and we can assume that the endpoint
does what it is intended to do (which is also indicated by the tests).

If anything gets added to the "user info" later, the use case can either be
renamed to reflect that, causing a cascade of necessary changes to its
consumers, or another use case can be added that is more general and provides
more "user info" besides just the full name.

This shows how changes to higher levels can have a cascading effect on lower
levels if we are not precise in our naming, and if we depend too much on
implementation details. Adding interfaces does not automatically decouple
(*Indirection Without Abstraction*); it is the *intentional ignorance of details*,
being *implementation-agnostic*, that is necessary to achieve real decoupling.
And sometimes an interface is not even required.