Make unit tests independent from real data
In the integration tests it is ok to use actual product IDs
from the database, but the unit tests should be independent
from the database and keep working even when the actual data
gets changed, e.g., when a product is renamed. So the unit
tests can provide their own hypothetical products for testing
the functionality.