Clean up brittle ArchUnit tests
Some of these tests were used as examples in the ArchUnit kata,
but some of the examples are not applicable in general and are
very brittle (i.e., they break easily for unintended reasons).
We remove some of the more brittle example tests and make sure
all tests are passing again. The examples are still available
if needed on the ArchUnit kata branch.